requirements.txt
fastapi
uvicorn
pandas
yfinance
pydantic
requests
google-generativeai
python-dotenv
sqlalchemy
